refactor: change directory of Sign Up form #23
