--- src/features/authentication/signup/ui/SignupForm/SignupForm.module.scss
@@ -1,3 +1,0 @@
-.signup_form {
-  font-size: 16px;
-}
