--- src/features/authentication/signup/ui/SignupForm/index.tsx
@@ -1,7 +1,0 @@
-import styles from './SignupForm.module.scss';
-
-function SignupForm(): JSX.Element {
-  return <div className={styles.signup_form}>SignupForm</div>;
-}
-
-export default SignupForm;
