+++ src/pages/index.ts
@@ -1,0 +1,4 @@
+import LoginPage from './login/LoginPage';
+import SignUpPage from './signUp/SignUpPage';
+
+export { LoginPage, SignUpPage };
