--- src/widgets/registration/index.ts
@@ -1,3 +1,0 @@
-import RegistrationForm from './ui/layout/RegistrationForm';
-
-export default RegistrationForm;
