--- src/widgets/registration/model/index.ts
@@ -1,2 +1,0 @@
-const temp = (): void => {};
-export default temp;
