--- src/widgets/registration/ui/email/RegistrationEmail.module.scss
@@ -1,3 +1,0 @@
-.registration_email {
-  font-size: 16px;
-}
