--- src/widgets/registration/ui/email/index.tsx
@@ -1,7 +1,0 @@
-import styles from './RegistrationEmail.module.scss';
-
-function RegistrationEmail(): JSX.Element {
-  return <div className={styles.registration_email}>Registration Email</div>;
-}
-
-export default RegistrationEmail;
