--- src/widgets/registration/ui/password/RegistrationPassword.module.scss
@@ -1,3 +1,0 @@
-.registration_password {
-  font-size: 16px;
-}
