--- src/widgets/registration/ui/password/index.tsx
@@ -1,7 +1,0 @@
-import styles from './RegistrationPassword.module.scss';
-
-function RegistrationPassword(): JSX.Element {
-  return <div className={styles.registration_password}>Registration Password</div>;
-}
-
-export default RegistrationPassword;
